fix(debug): change sups type from u32 to u64
- Updated the `sups` parameter in the `render_debug_overlay` function
  to use `u64` instead of `u32` for better handling of larger values.

+++ src/debug.rs
@@ -7,7 +7,7 @@ use sdl2::video::Window;
 pub fn render_debug_overlay(
     canvas: &mut Canvas<Window>,
     tiles_texture: &mut Texture<'_>,
-    sups: u32,
+    sups: u64,
     fps: u32,
     zoom: f64,
 ) {
